requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pydantic-settings==2.11.0
eval-type-backport==0.3.1
SQLAlchemy==2.0.48
alembic==1.16.5
psycopg[binary]==3.2.13
email-validator==2.3.0
